Miscellaneous call cleanups.
- I deleted fileresolver.go. It was unused.
- I fixed a type in balancer.go.
- I removed a stale TODO in call.go.
- I moved the code in constantresolver.go to resolver.go.
