+++ internal/net/call/resolver.go
@@ -16,6 +16,7 @@ package call
 
 import (
 	"context"
+	"fmt"
 )
 
 // A Resolver manages a potentially changing set of endpoints. For example:
@@ -110,3 +111,30 @@ type Version struct {
 // Should we move the version into a common file? Right now we have a duplicate
 // version struct that is used both by the gke/store and stub/resolver.
 var Missing = Version{"__tombstone__"}
+
+// constantResolver is a trivial constant resolver that returns a fixed set of
+// endpoints.
+type constantResolver struct {
+	endpoints []Endpoint
+}
+
+var _ Resolver = &constantResolver{}
+
+// NewConstantResolver returns a new resolver that returns the provided
+// set of endpoints.
+func NewConstantResolver(endpoints ...Endpoint) Resolver {
+	return &constantResolver{endpoints: endpoints}
+}
+
+// IsConstant implements the Resolver interface.
+func (*constantResolver) IsConstant() bool {
+	return true
+}
+
+// Resolve implements the Resolver interface.
+func (c *constantResolver) Resolve(_ context.Context, version *Version) ([]Endpoint, *Version, error) {
+	if version != nil {
+		return nil, nil, fmt.Errorf("unexpected non-nil version %v", *version)
+	}
+	return c.endpoints, nil, nil
+}
